Unifies `DeltaBalance` and `SignedAmount`
