+++ crates/namada/src/ledger/native_vp/ethereum_bridge/bridge_pool_vp.rs
@@ -32,7 +32,7 @@ use namada_tx::Tx;
 use crate::address::{Address, InternalAddress};
 use crate::eth_bridge_pool::{PendingTransfer, TransferToEthereumKind};
 use crate::ethereum_events::EthAddress;
-use crate::ledger::native_vp::{Ctx, NativeVp, StorageReader};
+use crate::ledger::native_vp::{Ctx, NativeVp, SignedAmount, StorageReader};
 use crate::storage::Key;
 use crate::token::storage_key::balance_key;
 use crate::token::Amount;
@@ -42,13 +42,6 @@ use crate::vm::WasmCacheAccess;
 #[error(transparent)]
 /// Generic error that may be returned by the validity predicate
 pub struct Error(#[from] eyre::Error);
-
-/// A positive or negative amount
-#[derive(Copy, Clone)]
-enum SignedAmount {
-    Positive(Amount),
-    Negative(Amount),
-}
 
 /// An [`Amount`] that has been updated with some delta value.
 #[derive(Copy, Clone)]
